==> 01_building_dominant_dm_drug.ipynb <==
import pandas as pd
import numpy as np

## Importing

dm_df = pd.read_pickle('./dm_drug_class_numbers_encr.pkl')

dm_df

## Building Dominant Drug

dm_df.columns[0:8]

handy = dm_df[
    ['A10A', 'A10BA', 'A10BB', 'A10BD', 'A10BF', 'A10BG', 'A10BJ', 'A10BX']
].copy()
handy

dominant_df = pd.DataFrame(index=dm_df.index)

# Determining insulin
dominant_df.loc[handy['A10A'] > 0, 'insulin'] = True

dominant_df['insulin'].value_counts(dropna=False, normalize=True)

dominant_df

handy.sum(axis=1)

temp = handy.apply(lambda c: c / handy.sum(axis=1).values)
temp

cumsum_df = temp.stack().to_frame().groupby('member').apply(lambda g: g.cumsum())

cumsum_df

### Sandbox

# Specifying dominant drugs
cut_off = 0.8
(cumsum_df > cut_off).head(20)

a = (cumsum_df > cut_off).loc[0]
b = (cumsum_df > cut_off).loc[0].shift(1)

a

b

temp = a & b

temp

' '.join(a.loc[~ temp.values].index.sort_values())

### main

def get_dominant_string(g):
    g = g.droplevel(0, axis=0)
    temp = g & g.shift()
    output = ' '.join(g.loc[~ temp.values].index.sort_values())
    # print(output)
    return output

dominants_dict = dict()

dominants_dict[cut_off] = (cumsum_df > cut_off).groupby(
    level=0
).apply(get_dominant_string)

dominants_dict[0.8]

dominant_df[cut_off] = dominants_dict[cut_off]

dominant_df

dominant_df.loc[dominant_df['insulin'] == True, 0.8].value_counts(normalize=True).head(20)

dominant_df.loc[dominant_df['insulin'].isna(), 0.8].value_counts(normalize=True).head(20)

dominant_df

# Saving
# dominant_df.to_pickle('./results/dominants_0.8.pkl')
