==> 01_person_table.ipynb <==
import pandas as pd
import numpy as np

province_id = pd.read_excel('../../diabeteschronicabuserprojectihio/Test_sample/Test/province_id.xlsx')
province_id.set_index('province_id_iso', inplace=True)
province_id = province_id['province_name_en']
province_id

# table = pd.read_csv('../person_table.csv', index_col=[0, 1])
table = pd.read_clipboard()
table

table.rename({'all': 'All'}, axis=1, inplace=True)

temp = table.loc['province'].index.astype(float)

prov_index = pd.MultiIndex.from_product([['province'], province_id.loc[temp].values])

prov_index

prov = pd.DataFrame(index=prov_index, data=table.loc['province'].values, columns=table.columns)

table = pd.concat(
    [table.drop('province', axis=0, level=0), prov]
)

table

# Saving person_table.xlsx
# table.to_excel('../person_table_edited.xlsx')
# table.to_excel('../Edited fund/person_table_f.xlsx')
