==> 01_preprocessing_data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

## Opening Files and Merging

prescription = pd.read_pickle('../nosakh_ARASH.pkl')
# Parsing dates
prescription['date'] = pd.to_datetime(prescription['date'])
prescription

prescription.info()

physician = pd.read_pickle('../physician.pkl')
physician

physician.info()

# Merging specialty_en to prescription
prescription_m_1 = prescription.merge(
    physician[['physician', 'specialty_en']],
    how='left',
    on='physician'
)
prescription_m_1

profile = pd.read_pickle('../profile_ARASH.pkl')
# Parsing birthdate
profile['birthdate'] = pd.to_datetime(profile['birthdate'])
profile

profile.info()

# Merging birthdate and gender to prescription
prescription_m_2 = prescription_m_1.merge(
    profile[['member', 'birthdate', 'gender']],
    how='left',
    on='member'
)
prescription_m_2

### Processing service before merging

service = pd.read_parquet('../service.parquet')
service

service.info()

# Labeling DM drugs and antibiotics in service for future merging with
# prescription_m_2
dm_atc_3 = ['A10A', 'A10B']
ab_atc_2 = ['J01', 'J02', 'J04', 'J05']
service['dm'] = service['atc_3'].isin(dm_atc_3)
service['ab'] = service['atc_2'].isin(ab_atc_2)

service

service.loc[service['dm']].shape

service.loc[service['ab']].shape

# Merging (atc, dm, ab) from service to prescription_m_2 on service
prescription_m_3 = prescription_m_2.merge(
    service[['service', 'atc', 'dm', 'ab']],
    how='left',
    on='service'
)
prescription_m_3

prescription_m_3.loc[prescription_m_3['dm'] | prescription_m_3['ab']]['member'].nunique()

absent_physician = ~ prescription_m_3['physician'].isin(physician['physician'])
absent_physician.sum()

prescription_m_3.loc[absent_physician, 'physician'].nunique()

There are more physicians in the prescription data than in the physician dataframe.

### Merging DDD data from ab_service_final

ab_service_final = pd.read_csv('../ab_service_final.csv', index_col=0)
ab_service_final['service'] = ab_service_final.index
ab_service_final

# Merging (calc_ddd, recheck) from ab_service_final to prescription_m_3
# on service
prescription_m_4 = prescription_m_3.merge(
    ab_service_final[['service', 'calc_ddd', 'recheck']],
    how='left',
    on='service'
)
prescription_m_4

temp = prescription_m_4['claimed'] == prescription_m_4['insurance_paid']

prescription_m_4.shape[0] - temp.sum()

The claimed and insurance_paid columns aren't the same.

prescription_m_4.columns

prescription_m_4['deduction'].nunique()

prescription_m_4.info()

### Removing unwanted columns

unwanted = ['physician', 'institute', 'claimed', 'fund']
prescription_m_final = prescription_m_4.drop(unwanted, axis=1)

prescription_m_final

prescription_m_final.info()

## Saving the final dataframe
It contains prescription data except ['physician', 'institute', 'claimed', 'fund'] columns. Plus (specialty_en) from physician and (birthdate, gender) from profile and (atc, dm, ab) from service and mask on that and (calc_ddd, recheck) from ab_service_final.

# prescription_m_final.to_csv('./prescription_merged_final.csv')

## Sandbox

from statistics import mode

temp = prescription.groupby(['member', 'prescription'])['province'].apply(lambda g: g.iloc[0])
temp = pd.DataFrame(temp)
temp

# temp['prescription'] = temp.index.get_level_values('prescription')
temp.groupby('member')['province'].apply(mode)

temp

bins = pd.IntervalIndex.from_tuples([(1, 100), (101, 532), (533, 1000)], 'both')
bins

pd.cut(prescription['quantity'], bins=bins)

prescription['quantity_cat'] = pd.cut(prescription['quantity'], bins=bins)
prescription['quantity_cat']

pd.IntervalIndex.from_tuples([
    (18, 24),
    (25, 34),
    (35, 44),
    (45, 54),
    (55, 64),
    (65, 74),
    (75, 95)
],
    'both'
)
