==> 01_starting_exploration.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

P = '1'
location = '..\\IHIO_share_A\\prescription\\part' + P + '.pkl'

part = pd.read_pickle(location)

part['date'] = pd.to_datetime(part['date'])

service = pd.read_pickle(r'../IHIO_share_A/service.pkl')
service.set_index('service', inplace=True)

dm_filt = service['atc_3'].isin(['A10A', 'A10B'])
dm_service_codes = service.loc[dm_filt].index
dm_service_codes.shape

part

is_dm_part = part['service'].isin(dm_service_codes)

part_dm = part.loc[part['member'].isin(part.loc[is_dm_part, 'member'].unique())]

part_dm

# Importing ab_service
ab_service = pd.read_csv('.\\ab_service_final.csv', index_col=0)

ddd = ab_service['calc_ddd']

## Calculating days of ab use for each member.

part_dm['member'].unique().shape

part['member'].unique().shape

# ab_mask is the services of part_dm that are ab
ab_mask = part_dm['service'].isin(ddd.index)
part_dm_ab = part_dm.loc[ab_mask]
part_dm_ab.loc[:, 'days'] = part_dm_ab['quantity'] * ddd[part_dm_ab['service']].values

part_dm_ab

part_dm_ab['member'].nunique()

ab_mask.sum()

part_dm_ab['days'].describe()

# Calculating sum of days for each member.
part_dm_ab.groupby('member')['days'].sum().describe()

# sns.histplot(x=np.log10(part_dm_ab['days']), bins=20)
# plt.gca().set_title('log10 days for members in part1')
# plt.xticks(np.arange(-2.5, 4, 0.5))

# The sum of days for each dm patient.
# sns.histplot(x=np.log10(part_dm_ab.groupby('member')['days'].sum()), bins=20)

# sns.kdeplot(x=np.log10(part_dm_ab['days']), bw_adjust=2)

# Non-dm patients
part_n = part.loc[~ part['member'].isin(part_dm['member'].unique())]

part_n

# Calculating the same statisitics for non-dm patients
# ab_mask_n is the services of part_n that are ab
ab_mask_n = part_n['service'].isin(ddd.index)
part_n_ab = part_n.loc[ab_mask_n]
part_n_ab.loc[:, 'days'] = part_n_ab['quantity'] * ddd[part_n_ab['service']].values

part_n_ab

part_n_ab['member'].nunique()

part_n_ab['days'].describe()

part_n_ab.groupby('member')['days'].sum().describe()

# sns.histplot(x=np.log10(part_n_ab['days']), bins=20)
# plt.gca().set_title('non-dm log10 days for members in part1')
# plt.xticks(np.arange(-2.5, 4, 0.5))

# The sum of days for each non-dm patient.
# sns.histplot(x=np.log10(part_n_ab.groupby('member')['days'].sum()), bins=20)

part['date'].agg([min, max])

fund_dm_weights = part_dm['fund'].value_counts(normalize=True).sort_index()
fund_dm_weights

part_n['fund'].value_counts(normalize=True).sort_index()

fund_cat = sorted(part_n['fund'].unique())
fund_weights = pd.Series(index=fund_cat, dtype=float)
fund_weights.loc[fund_dm_weights.index] = fund_dm_weights
fund_weights

test_weights = fund_weights.loc[part_n['fund']]

df = part_n[['fund', 'prescription']]
df['fund'].value_counts(normalize=True)

df['weight'] = test_weights.values
df

sam = df.sample(n=1000000, weights='weight')

sam['fund'].value_counts(normalize=True).sort_index()
