==> 02_building_drug_class_table.ipynb <==
import pandas as pd
import numpy as np
from statsmodels.stats.proportion import proportion_confint

# DM

## Importing

dm_df = pd.read_pickle('./dm_drug_class_numbers_encr.pkl')
dm_df

## Building dm drug-class Table

dm_classes = ['A10A', 'A10BA', 'A10BB', 'A10BD', 'A10BF', 'A10BG', 'A10BJ', 'A10BX']

dm_table = pd.DataFrame(columns=sorted(dm_df['cat'].unique()))

dm_table.loc['total'] = dm_df.groupby('cat')[dm_classes].sum().T.sum()

dm_table

dm_table = pd.concat(
    [dm_table, dm_df.groupby('cat')[dm_classes].sum().T]
)

dm_concat_dict = dict()

dm_concat_dict['numbers'] = dm_table

dm_concat_dict['proportions'] = dm_table.loc[dm_classes] / dm_table.loc['total']

pd.concat(dm_concat_dict, axis=1)

dm_concat_dict['ci_low'], dm_concat_dict['ci_up'] = proportion_confint(
    dm_table.loc[dm_classes],
    dm_table.loc['total']
)

pd.concat(dm_concat_dict, axis=1)

## Saving dm_drug_class_table

# Saving dm_drug_class_table
# pd.concat(dm_concat_dict, axis=1).to_excel('./results/dm_drug_class_table.xlsx')

# AB

## Importing

ab_df = pd.read_pickle('./ab_drug_class_numbers_encr.pkl')
ab_df.drop(ab_df.loc[ab_df['cat'] == 0].index, axis=0, inplace=True)
ab_df

## Building ab drug-class Table

ab_classes = ['J01A', 'J01B', 'J01C', 'J01D', 'J01E', 'J01F', 'J01G', 'J01M', 'J01X']

ab_table = pd.DataFrame(columns=sorted(ab_df['cat'].unique()))

ab_table.loc['total_service'] = ab_df.groupby('cat')[ab_classes].sum().T.sum()

ab_table.loc['total_person'] = ab_df['cat'].value_counts()

ab_table

ab_table = pd.concat(
        [ab_table, ab_df.groupby('cat')[ab_classes].sum().T]
)

ab_concat_dict = dict()

ab_concat_dict['numbers'] = ab_table

ab_concat_dict['number/total_service'] = ab_table.loc[ab_classes] / ab_table.loc['total_service']

ab_concat_dict['s_ci_low'], ab_concat_dict['s_ci_up'] = proportion_confint(
    ab_table.loc[ab_classes],
    ab_table.loc['total_service']
)
# Changing column names
ab_concat_dict['s_ci_low'].columns = ab_table.columns
ab_concat_dict['s_ci_up'].columns = ab_table.columns

ab_concat_dict['number/person'] = (
    ab_table.loc[ab_classes] / ab_table.loc['total_person']
)

ab_concat_dict['p_ci_low'], ab_concat_dict['p_ci_up'] = proportion_confint(
    ab_table.loc[ab_classes],
    ab_table.loc['total_person']
)
# Changing column names
ab_concat_dict['p_ci_low'].columns = ab_table.columns
ab_concat_dict['p_ci_up'].columns = ab_table.columns

pd.concat(ab_concat_dict,  axis=1)

pd.concat(ab_concat_dict, axis=1)

temp = pd.DataFrame(columns=sorted(ab_df['cat'].unique()))
for cat in sorted(ab_df['cat'].unique()):
    temp_cat = ab_df.loc[ab_df['cat'] == cat]
    temp[cat] = temp_cat[ab_classes].isna().sum() / temp_cat.shape[0]

ab_concat_dict['ever_proportion'] = temp

pd.concat(ab_concat_dict, axis=1)

## Saving ab_drug_class_table

# Saving ab_drug_class_table
# pd.concat(ab_concat_dict, axis=1).to_csv('./results/ab_drug_class_table_02.csv')

# Dominant Drug

## Importing

dominant_df = pd.read_pickle('./results/dominants_0.8.pkl')
# By checking the results, we found out we don't need to consider one-time use
# of insulin.
dominant_df = dominant_df[0.8].to_frame('drug')
dominant_df

np.ones(dominant_df['drug'].nunique()) * dominant_df['drug'].shape[0]

## Building Dominant Table

def my_proportion_confint(df, col):
    low, up = proportion_confint(
        df['drug'].value_counts(),
        df.shape[0]
    )
    return low, up

dominant_df['cat'] = dm_df['cat']

dominant_df['cat'].value_counts()

major_classes = [
    'A10BB',
    'A10A A10BA',
    'A10BA A10BB A10BG',
    'A10BA A10BG',
    'A10BG',
    'A10BA',
    'A10A',
    'A10A A10BA A10BB',
    'A10BA A10BB',
    'A10BA A10BB A10BF'
 ]

dominant_df.loc[~ dominant_df['drug'].isin(major_classes), 'drug'] = 'other'

dominant_proportions = dominant_df.groupby('cat')['drug'].value_counts(normalize=True)
dominant_proportions

dominant_proportions[0].sort_values(ascending=False).to_frame('proportion').index

dominant_dict = dict()

for cat in sorted(dm_df['cat'].unique()):
    # Proportion
    # This also drops na so I can use the shape for confint
    temp_cat = dominant_proportions[cat].sort_values(ascending=False).to_frame(
        'proportion')
    # confidence intervals
    temp_cat['ci_low'], temp_cat['ci_up'] = my_proportion_confint(
        dominant_df.loc[dominant_df['cat'] == cat, ['drug']],
        'drug'
    )
    # Adding drug as a column
    temp_cat.reset_index(inplace=True)
    # Cumulative proportion sum
    temp_cat['cumulative'] = temp_cat['proportion'].cumsum().values
    dominant_dict[cat] = temp_cat

pd.concat(dominant_dict, axis=1)

## Saving dominant_dm_class

# Saving dominant_dm_class
# pd.concat(dominant_dict, axis=1).to_excel('./results/dominant_dm_class_table.xlsx')

# Saving dominant_dm_class_edited
# pd.concat(dominant_dict, axis=1).to_excel('../Results to Edit/dominant_dm_class_edited.xlsx')
