==> 02_prescription_table.ipynb <==
import pandas as pd
import numpy as np
from statsmodels.stats.proportion import proportion_confint

table = pd.read_excel('../complete_prescription_table.xlsx', header=[0, 1],
                      index_col=0)
table = table['number']
# Removing outliers so No AB would be zero.
table.loc['ab', '0'] = 0
# Renaming
table.rename({'all': 'All', '0': 'No Antibiotic', '1':'Q1', '2': 'Q2', 
              '3': 'Q3', '4': 'Q4'},
              axis=1, inplace=True)
#
table

concat_dict = dict()

concat_dict['number'] = table
concat_dict['proportion'] = table.loc[['ab', 'dm']] / table.loc['total']

concat_dict['proportion']

concat_dict['low'], concat_dict['up'] = proportion_confint(
    table.loc[['ab', 'dm']], table.loc['total']
)

col_names = table.columns.to_list()

def rename_cols(df, col_names):
    df.columns = col_names
    return df

concat_dict['low'] = rename_cols(concat_dict['low'], col_names=col_names)
concat_dict['up'] = rename_cols(concat_dict['up'], col_names=col_names)

concat_dict['low']
concat_dict['up']

pd.concat(concat_dict, axis=1)

def build_string(val_df, low_df, up_df, to_percentage, decimals=2):

    if to_percentage:
        val_df = val_df * 100
        low_df = low_df * 100
        up_df = up_df * 100

    val_df = val_df.map(lambda x: f'{x:.2f}')
    low_df = low_df.map(lambda x: f'{x:.2f}')
    up_df = up_df.map(lambda x: f'{x:.2f}')



    return val_df + ' (' + low_df + '-' + up_df + ')'

temp = build_string(
    concat_dict['proportion'].loc[['ab', 'dm']],
    concat_dict['low'].loc[['ab', 'dm']],
    concat_dict['up'].loc[['ab', 'dm']],
    to_percentage=True
)

temp.loc['total'] = table.loc['total']

temp

# Saving prescription_table_edited.xlsx
# temp.to_excel('../prescription_table_edited.xlsx')
