==> 03_ab_drug_class_table.ipynb <==
import pandas as pd
import numpy as np
from statsmodels.stats.proportion import proportion_confint

prev_df = pd.read_excel('../ab_drug_class_table_02.xlsx', header=[0, 1],
                        index_col=0)
prev_df

def build_string(val_df, low_df, up_df, to_percentage, decimals=2):

    if to_percentage:
        val_df = val_df * 100
        low_df = low_df * 100
        up_df = up_df * 100

    val_df = val_df.map(lambda x: f'{x:.2f}')
    low_df = low_df.map(lambda x: f'{x:.2f}')
    up_df = up_df.map(lambda x: f'{x:.2f}')



    return val_df + ' (' + low_df + '-' + up_df + ')'

proportion_df = prev_df['number/total_service']

drug_classes = ['J01A', 'J01B', 'J01C', 'J01D', 'J01E', 'J01F', 'J01G', 'J01M', 'J01X']

proportion_df = proportion_df.loc[drug_classes]
proportion_df

concat_dict = dict()

concat_dict['proportion'] = build_string(
    prev_df.loc[drug_classes, 'number/total_service'],
    prev_df.loc[drug_classes, 's_ci_low'],
    prev_df.loc[drug_classes, 's_ci_up'],
    to_percentage=True,
    decimals=2
)

concat_dict['proportion']

### Building everproportion and CI

ab_df = pd.read_pickle('../../Drug class data/ab_drug_class_numbers_encr.pkl')
ab_df.drop(ab_df.loc[ab_df['cat'] == 0].index, axis=0, inplace=True)
ab_df

temp = pd.DataFrame(columns=sorted(ab_df['cat'].unique()))
for cat in sorted(ab_df['cat'].unique()):
    temp_cat = ab_df.loc[ab_df['cat'] == cat]
    temp[cat] = (temp_cat[drug_classes] > 0).sum()

temp

temp_low, temp_up = proportion_confint(
    temp,
    prev_df.loc['total_person', 'numbers']
)
temp_low.columns = temp.columns
temp_up.columns = temp.columns

temp_ever = pd.DataFrame(columns=sorted(ab_df['cat'].unique()))
for cat in sorted(ab_df['cat'].unique()):
    temp_cat = ab_df.loc[ab_df['cat'] == cat]
    temp_ever[cat] = (temp_cat[drug_classes] > 0).sum() / temp_cat.shape[0]

concat_dict['ever_proportion'] = build_string(
    temp_ever,
    temp_low,
    temp_up,
    to_percentage=True,
    decimals=2
)

concat_dict['proportion'].loc['total'] = prev_df.loc['total_service', 'numbers']
concat_dict['proportion']

total_person = ab_df['cat'].value_counts().sort_index()

concat_dict['ever_proportion'].loc['total'] = total_person
concat_dict['ever_proportion']

col_name_mapper = {1:'Q1', 2: 'Q2', 3: 'Q3', 4: 'Q4'}

final_df = pd.concat(concat_dict, axis=1).rename(col_name_mapper, axis=1, level=1)

final_df

# Saving ab_drug_class_table_edited.xlsx
# final_df.to_excel('../ab_drug_class_table_edited.xlsx')
