==> 03_building_costs_table.ipynb <==
import pandas as pd
import numpy as np
import statsmodels.stats.api as sm
from itertools import product

## Importing Data

costs = pd.read_pickle('../1403-4-16 export/yearly_costs_encr.pkl')

# data = costs['data']
data = pd.read_pickle('../edited_fund_person_data.pkl')
data

costs.drop('data', axis=1, level=0, inplace=True)
costs.fillna(0, inplace=True)
costs.rename({'pocket_adj': 'pocket', 'insurance_adj': 'insurance'},
             axis=1, level=1, inplace=True)
costs

# Setting zero category ab_costs to zero.
costs.loc[data['cat'] == 0, 'ab'] = 0

## Building Table

concat_dict = dict()

drugs = ['ab', 'dm', 'other']
bag = ['pocket', 'insurance']

for d in drugs:
    costs[(d, 'sum')] = costs[d].sum(axis=1)
costs[('total', 'total')] = costs[product(drugs, ['sum'])].sum(axis=1)

costs.sort_index(axis=1, inplace=True)
costs

# costs = costs.stack([0, 1])

costs

for col in data.columns.drop('cat'):
    concat_dict[col] = costs.groupby([data['cat'], data[col]])[costs.columns].mean()
df = pd.concat(concat_dict)

concat_dict = dict()
for col in data.columns:
    concat_dict[col] = costs.groupby(data[col])[costs.columns].mean()
df_all_cat = pd.concat(concat_dict)

df_all_cat

## Computing Confidence Intervals

import scipy.stats as st


def get_ci(g):
    low, up = st.t.interval(confidence=0.95, df=g.shape[0]-1,
                            loc=np.mean(g, axis=0),
                            scale=st.sem(g))
    return low, up

concat_dict = dict()
for col in data.columns:
    concat_dict[col] = costs.groupby(data[col])[costs.columns].apply(lambda g: get_ci(g))

concat_dict['gender']

df_all_cat_low = pd.DataFrame(index=df_all_cat.index, columns=df_all_cat.columns)
df_all_cat_up = pd.DataFrame(index=df_all_cat.index, columns=df_all_cat.columns)

for idx in df_all_cat.index.to_flat_index():
    col, col_val = idx
    # low
    df_all_cat_low.loc[(col, col_val)] = concat_dict[col][col_val][0]
    # up
    df_all_cat_up.loc[(col, col_val)] = concat_dict[col][col_val][1]

df_all_cat_low

df_all_cat_up

# # Saving costs by each variable values.
# pd.concat(
#     [df_all_cat, df_all_cat_low, df_all_cat_up],
#     axis=1,
#     keys=['val', 'low', 'up']
# ).to_excel('./results/costs_by_variables_mean_ci_f.xlsx')

df

concat_dict = dict()

concat_dict_2 = dict()

concat_dict_2['val'] = df
concat_dict_2['low'] = pd.DataFrame(index=df.index, columns=df.columns)
concat_dict_2['up'] = pd.DataFrame(index=df.index, columns=df.columns)

for idx in df.index.to_flat_index():
    col, cat, col_val = idx
    # low
    # concat_dict_2['low'].loc[idx] = 

# Saving costs by cat and other variables.
# df.to_excel('./results/costs_by_cat_and_variables.xlsx')

## Computing median iqr

costs

concat_dict = dict()
for col in data.columns.drop('cat'):
    concat_dict[col] = costs.groupby([data['cat'], data[col]])[costs.columns].median()
df_cat_median = pd.concat(concat_dict)

df_cat_median

# Saving median costs by cat and other variables.
# df_cat_median.to_excel('./results/costs_by_cat_and_variables_median.xlsx')

concat_dict = dict()
concat_dict_2 = dict()
for q in [0.25, 0.5, 0.75]:
    for col in data.columns:
        concat_dict[col] = costs.groupby(data[col])[costs.columns].quantile(q)
    concat_dict_2[q] = pd.concat(concat_dict)

# Saving quartiles of price by variables.
# pd.concat(concat_dict_2, axis=1).to_excel('./results/costs_by_variables_quartiles_f.xlsx')
