==> 04_dm_drug_class_table.ipynb <==
import pandas as pd
import numpy as np
from statsmodels.stats.proportion import proportion_confint

prev_df = pd.read_excel('../dm_drug_class_table.xlsx', header=[0, 1], index_col=0)
prev_df

def build_string(val_df, low_df, up_df, to_percentage, decimals=2):

    if to_percentage:
        val_df = val_df * 100
        low_df = low_df * 100
        up_df = up_df * 100

    val_df = val_df.map(lambda x: f'{x:.2f}')
    low_df = low_df.map(lambda x: f'{x:.2f}')
    up_df = up_df.map(lambda x: f'{x:.2f}')



    return val_df + ' (' + low_df + '-' + up_df + ')'

drug_classes = ['A10A', 'A10BA', 'A10BB', 'A10BD', 'A10BF', 'A10BG', 'A10BJ', 'A10BX']

concat_dict = dict()

concat_dict['proportion'] = build_string(
    prev_df.loc[drug_classes, 'proportions'],
    prev_df.loc[drug_classes, 'ci_low'],
    prev_df.loc[drug_classes, 'ci_up'],
    to_percentage=True,
    decimals=2
)

### Building everproportion and CI

dm_df = pd.read_pickle('../../Drug class data/dm_drug_class_numbers_encr.pkl')
dm_df

temp = pd.DataFrame(columns=sorted(dm_df['cat'].unique()))
for cat in sorted(dm_df['cat'].unique()):
    temp_cat = dm_df.loc[dm_df['cat'] == cat]
    temp[cat] = (temp_cat[drug_classes] > 0).sum()

temp

nobs = dm_df['cat'].value_counts().sort_index()

def my_proportion_confint(count, nobs, col_names):
    
    low, up = proportion_confint(count, nobs)

    low.columns = col_names
    up.columns = col_names
    
    return low, up

temp_low, temp_up = my_proportion_confint(temp, nobs, temp.columns)

temp_low

temp_up

temp_ever = pd.DataFrame(columns=sorted(dm_df['cat'].unique()))
for cat in sorted(dm_df['cat'].unique()):
    temp_cat = dm_df.loc[dm_df['cat'] == cat]
    temp_ever[cat] = (temp_cat[drug_classes] > 0).sum() / temp_cat.shape[0]

temp_ever

concat_dict['ever_proportion'] = build_string(
    temp_ever,
    temp_low,
    temp_up,
    to_percentage=True,
    decimals=2
)

concat_dict['proportion'].loc['total'] = prev_df.loc['total', 'numbers']
concat_dict['proportion']

total_person = dm_df['cat'].value_counts().sort_index()
total_person

concat_dict['ever_proportion'].loc['total'] = total_person
concat_dict['ever_proportion']

col_name_mapper = {0: 'No Antibiotic', 1:'Q1', 2: 'Q2', 3: 'Q3', 4: 'Q4'}

final_df = pd.concat(concat_dict, axis=1).rename(col_name_mapper, axis=1, level=1)

final_df

# Saving dm_drug_class_table_edited.xlsx
# final_df.to_excel('../dm_drug_class_table_edited.xlsx')
