==> 04_plotting_trend.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_csv('../1403-4-16 export/trend_service_numbers.csv')
df

drug_class = ['J01A', 'J01B', 'J01C', 'J01D', 'J01E', 'J01F', 'J01G', 'J01M', 'J01X']

fig, axes = plt.subplots(2, 2, figsize=(10, 10), dpi=300, layout='constrained')

for cat, ax in zip(sorted(df['cat'].unique()), axes.flatten()):
    for d in drug_class:
        sns.lineplot(df.loc[df['cat'] == cat], x='season', y=d, ax=ax, label=d)
    ax.set_ylabel('')
    ax.set_title(f'{cat=}')
