==> 04_sampling_case_matched.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path
import os

part = pd.read_pickle(r'C:\Users\a.bagherian\Desktop\working\Diabetics Combined\part2_dm_combined.pkl')
age = pd.read_csv(r'C:\Users\a.bagherian\Desktop\working\Ages\part2_dm_age.csv', index_col='member')

part

age

profile = pd.read_pickle(r'C:\Users\a.bagherian\Desktop\IHIO_share_A\profile.pkl')
profile

1 for male\
0 for female

age['gender'] = profile.set_index('member').loc[age.index, 'gender']

age

age['age'].describe()

# Cleaning age a bit.
age['age'] = np.where(age['age'] < 0, 0, age['age'])
age['age'] = np.where(age['age'] > 100, 100, age['age'])

age['age'].describe()

n_categories = 20

age['age_cat'] = pd.cut(age['age'], n_categories)

age['age_gender_cat'] = age['age_cat'].astype(str) + ' ' + age['gender'].astype(str)

age

n_weights = age['age_gender_cat'].value_counts()
n_weights.sort_index()

na_member = age.loc[age['age_cat'].isna()].index

## Importing non-dm data for sampling

non_dm = pd.read_pickle(r'C:\Users\a.bagherian\Desktop\IHIO_share_A\prescription\part2.pkl')
non_dm

non_dm = non_dm.drop(part.index, axis=0)
non_dm

non_dm = non_dm.merge(
    profile[['member', 'birthdate', 'gender']],
    how='left',
    on='member'
)

non_dm['date'] = pd.to_datetime(non_dm['date'])

non_dm['birthdate'] = pd.to_datetime(non_dm['birthdate'])

from statistics import mode


non_dm['age'] = (non_dm['date'] - non_dm['birthdate']) / pd.Timedelta(days=365.2425)
non_dm_ages = non_dm.groupby('member')['age'].apply(mode)

non_dm_ages.to_csv(r'C:\Users\a.bagherian\Desktop\working\Ages\part2_non_dm_age.csv')

# non_dm_ages = pd.read_csv(r'C:\Users\a.bagherian\Desktop\working\Ages\part2_non_dm_age.csv')

non_dm_ages

## Categorizing non_dm data

non_dm_ages = non_dm_ages.to_frame()
non_dm_ages = non_dm_ages.merge(
    profile[['member', 'gender']],
    how='left',
    on='member'
)

non_dm_ages

# Cleaning age a bit.
non_dm_ages['age'] = np.where(non_dm_ages['age'] < 0, 0, non_dm_ages['age'])
non_dm_ages['age'] = np.where(non_dm_ages['age'] > 100, 100, non_dm_ages['age'])

non_dm_ages['age'].describe()

non_dm_ages['age_cat'] = pd.cut(non_dm_ages['age'], n_categories)

non_dm_ages['age_gender_cat'] = non_dm_ages['age_cat'].astype(str) + ' ' + non_dm_ages['gender'].astype(str)

non_dm_ages

## Sampling

n_weights.to_frame()

non_dm_weights = n_weights.loc[non_dm_ages['age_gender_cat']]

non_dm_ages['n_weights'] = n_weights.loc[non_dm_ages['age_gender_cat']].values

non_dm_ages

def sampler(g, random_state=0):
    n = g['n_weights'].iloc[0]
    return g.sample(n=n, random_state=random_state)

matched_sample = non_dm_ages.groupby('age_gender_cat', group_keys=False).apply(sampler)

matched_sample

matched_sample['age_gender_cat'].value_counts().sort_index()

matched_sample['age'].describe()

matched_sample.to_pickle(r'C:\Users\a.bagherian\Desktop\working\part2_matched_sample.pkl')
