==> 05_computing_ci_prescription_table.ipynb <==
import pandas as pd
import numpy as np
from statsmodels.stats.proportion import proportion_confint

df = pd.read_csv('../1403-4-16 export/prescription_table_numbers.csv', index_col=0)

df

concat_dict = dict()
concat_dict['number'] = df
concat_dict['proportion'] = df.loc[['ab', 'dm']] / df.loc['total']

concat_dict['low'], concat_dict['up'] = proportion_confint(df.loc[['ab', 'dm']], df.loc['total'])

concat_dict['proportion']

# Saving prescription_table
# pd.concat(concat_dict, axis=1).to_excel('../19 Tir to send/complete_prescription_table.xlsx')
