==> 06_applying_age_filter.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
from pathlib import Path

Ages = Path(r'C:\Users\a.bagherian\Desktop\working\Ages\dm')

def remove_ages(age_loc, lower=18, upper=95):
    
    # Importing
    age = pd.read_csv(age_loc, index_col='member')

    # Rounding
    age = age.round()

    # Removing unwanted ages and missing ages.
    print(f'{lower=}, {upper=}')
    to_delete = ((age < lower) | (age > upper)).values
    to_delete = age.loc[to_delete].index
    age = age.drop(age.loc[age.isna().values].index, axis=0)
    age = age.drop(to_delete, axis=0)


    return age

for age_loc in os.listdir(Ages):
    age_loc = Ages / age_loc
    age = remove_ages(age_loc)
    loc = Ages / (age_loc.stem + '_18_95.csv')
    age.to_csv(loc)

## Removing unwanted ages from the whole data

Diabetics_Combined = Path(r'C:\Users\a.bagherian\Desktop\working\Diabetics Combined')
os.listdir(Diabetics_Combined)

def remove_ages_data(part_loc, age_loc, lower=18, upper=95):
    
    # Importing
    part = pd.read_pickle(part_loc)
    age = pd.read_csv(age_loc, index_col='member')

    # Removing
    print(f'{lower=}, {upper=}')
    want = age.index
    to_delete = part.loc[~ part['member'].isin(want)].index
    part = part.drop(to_delete, axis=0)



    return part

for part_loc, age_loc in zip(os.listdir(Diabetics_Combined), os.listdir(Ages_18_95)):

    age_loc = Ages_18_95 / age_loc

    part_loc = Diabetics_Combined / part_loc
    destination = Diabetics_Combined_18_95 / (part_loc.stem + '_18_95.pkl')
    part = remove_ages_data(part_loc, age_loc)
    part.to_pickle(destination)

age = pd.read_csv(age_loc, index_col='member')
age.loc[part['member'].unique()].describe()
