==> 07_determining_cut_offs.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
from pathlib import Path

parts_path = Path(r'C:\Users\a.bagherian\Desktop\working\Diabetics Combined 18_95')

destination_path = Path(r'C:\Users\a.bagherian\Desktop\working\one_df_days')

## Bulding one_day_df

# cols = ['member', 'service', 'prescription', 'days']

# df = pd.DataFrame(columns=cols)
# df

# df = pd.DataFrame(columns=['days', 'part'])
# df

# for i, pickle_name in enumerate(os.listdir(parts_path)):
#     part_number = i + 1
#     pickle_path = parts_path / pickle_name
    
#     # Importing
#     part = pd.read_pickle(pickle_path)

#     # # Filtering for ab rows
#     # part = part.loc[part['ab']]

#     # # Filtering for columns
#     # part = part[cols]

#     # computing member_days
#     member_days = pd.DataFrame(part.groupby('member')['days'].sum())
#     member_days['part'] = part_number

#     df = pd.concat(
#         [df, member_days],
#         axis=0
#     )

member_days_path = destination_path / 'member_days_18_95.pickle'
# df.to_pickle(member_days_path)

## _

member_days = pd.read_pickle(member_days_path)
member_days

# This file contains the days of each ab service for all diabetics. so the members of
# dm that didn't receive ab aren't here.
one_df_days_path = destination_path / 'one_df_days_18_95.pickle'
one_df_days = pd.read_pickle(one_df_days_path)

one_df_days

# Bye prescription
one_df_days['days'].describe()

# By member
member_days.describe() / 3

member_days.loc[member_days['days'] > pd.Timedelta(0, 'day')].describe()

## Looking for outliers

one_df_days['days'].quantile(0.99)

one_df_days.loc[one_df_days['days'] > one_df_days['days'].quantile(0.99)].describe()

limit = pd.Timedelta(30, 'day')
print(limit)
outliers = one_df_days.loc[one_df_days['days'] > limit]
outliers

outliers['part'] = member_days.loc[outliers['member'], 'part'].values
outliers

# Saving outlier prescriptions
# outliers.to_pickle(r'C:\Users\a.bagherian\Desktop\working\one_df_days\outlier_ab_prescriptions.pkl')

## Checking if the reason of outliers are ddds

ab_service = pd.read_csv(r'C:\Users\a.bagherian\Desktop\working\ab_service_final.csv',
                         index_col=0)
ab_service

temp = ab_service.loc[outliers['service'].unique()]
out_serv = temp.loc[temp['recheck'] == 1].index

out_serv

outliers.loc[outliers['service'].isin(out_serv)]

## _

## Removing outlier ab prescriptions and computing the member_days again

# df = pd.DataFrame(columns=['days', 'part'])
# df

# for i, pickle_name in enumerate(os.listdir(parts_path)):
#     part_number = i + 1
#     pickle_path = parts_path / pickle_name
    
#     # Importing
#     part = pd.read_pickle(pickle_path)

#     # Removing outliers
#     part = part.drop(outliers.loc[outliers['part'] == part_number].index, axis=0)

#     # # Filtering for ab rows
#     # part = part.loc[part['ab']]

#     # # Filtering for columns
#     # part = part[cols]

#     # computing member_days
#     member_days = pd.DataFrame(part.groupby('member')['days'].sum())
#     member_days['part'] = part_number

#     df = pd.concat(
#         [df, member_days],
#         axis=0
#     )

member_days_outliers_removed_path = destination_path / 'member_days_18_95_outliers_removed.pickle'
# df.to_pickle(member_days_outliers_removed_path)

days_final = pd.read_pickle(member_days_outliers_removed_path)
days_final

days_final.loc[days_final['days'] > pd.Timedelta(0, 'day')].describe() / 3

num_days = (days_final['days'] / pd.Timedelta(1, 'day')) / 3
num_days 

days_final.loc[num_days == 0, 'cat'] = 0
days_final.loc[(0 < num_days) & (num_days <= 5), 'cat'] = 1
days_final.loc[(5 < num_days) & (num_days <= 11), 'cat'] = 2
days_final.loc[(11 < num_days) & (num_days <= 21), 'cat'] = 3
days_final.loc[(21 < num_days), 'cat'] = 4

days_final['cat'].value_counts(normalize=True, dropna=False)

days_final

# Saving
# The member_cat.pickle file contains categories based on ab with outliers removed. and is 18_95. The index is member.
# days_final.to_pickle(destination_path / 'member_cat.pkl')
