==> 08_building_person_dataframe.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path
import os

## Importing Data

member_cat = pd.read_pickle(r'C:\Users\a.bagherian\Desktop\working\one_df_days\member_cat.pkl')
member_cat

profile = pd.read_pickle(r'C:\Users\a.bagherian\Desktop\IHIO_share_A\profile.pkl')
profile = profile[['member', 'gender', 'fund']].set_index('member')
profile = profile.loc[member_cat.index]
profile

## Buidling person_df

person_df = profile.copy()
person_df['cat'] = member_cat['cat']
person_df['part'] = member_cat['part']
person_df

person_df.index.name = 'member'

person_df['member'] = person_df.index
person_df

ages_path = Path(r'C:\Users\a.bagherian\Desktop\working\Ages\dm_18_95')
age_path_list = os.listdir(ages_path)
age_path_list

for age_path in age_path_list:

    age_df = pd.read_csv(ages_path / age_path, index_col='member')

    person_df.loc[age_df.index, 'age'] = age_df['age']

person_df

### Determining the province for each member

prescriptions_path = Path(r'C:\Users\a.bagherian\Desktop\working\Diabetics Combined 18_95')
prescriptions_path_list = os.listdir(prescriptions_path)
prescriptions_path_list

from statistics import mode

for part_loc in prescriptions_path_list:
    print('part', part_loc)
    part = pd.read_pickle(prescriptions_path / part_loc)
    part = part[['member', 'province', 'prescription']]
    provinces = part.groupby(['member', 'prescription'])['province'].apply(lambda g: g.iloc[0])
    provinces = pd.DataFrame(provinces)
    provinces = provinces.groupby('member')['province'].apply(mode)
    
    person_df.loc[provinces.index, 'province'] = provinces

person_df

# person_df check_point_1
destination = Path(r'C:\Users\a.bagherian\Desktop\working\one_df_days')

# person_df.to_pickle(destination / 'person_df_checkpoint_1.pkl')

### _

### Building age_cat

bins = pd.IntervalIndex.from_tuples(
    [
        (18, 24),
        (25, 34),
        (35, 44),
        (45, 54),
        (55, 64),
        (65, 74),
        (75, 95)
    ],
    'both'
)

person_df['age_cat'] = pd.cut(person_df['age'], bins=bins)

# person_df checkpoint_2
destination = Path(r'C:\Users\a.bagherian\Desktop\working\one_df_days')

# person_df.to_pickle(destination / 'person_df_checkpoint_2.pkl')

person_df['age_cat'].value_counts(normalize=True)

### _

province_id = pd.read_excel(r'C:\Users\a.bagherian\Desktop\IHIO_share_A\Others\Provinces and city\province_id.xlsx').set_index('province_id_ihio')
province_id.head()

person_df

## Changing age_cat to three groups

prev_df = pd.read_pickle(destination / 'person_df_checkpoint_2_final.pkl')
prev_df

new_df = prev_df.drop('age_cat', axis=1)
new_df

# Building new age_cat
new_bins = pd.IntervalIndex.from_tuples(
    [
        (18, 39),
        (40, 64),
        (65, 95)
    ],
    'both'
)

new_df['age_cat'] = pd.cut(new_df['age'], bins=new_bins)

new_df

# new_df['age_cat'].value_counts(normalize=True)

# Saving new_df with three age_groups

# new_df.to_pickle(r'E:\after_session_03_22\data\person_df_3agegroups.pkl')
