==> 08_dominant_dm_class_stackedbarplot.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_excel('../dominant_dm_class_edited.xlsx', header=[0, 1], index_col=0)
df

df

concat_dict = dict()
drug_order = df[(0, 'drug')].values

for cat in [0, 1, 2, 3, 4]:

    concat_dict[cat] = df[(cat)].set_index('drug').loc[drug_order, 'proportion']

plotting_df = pd.concat(concat_dict, axis=1)

plotting_df = plotting_df * 100
plotting_df = plotting_df.T

fig, ax = plt.subplots(1, 1, figsize=(8, 8), dpi=300, layout='constrained')
fig.set_facecolor('white')

plotting_df.plot.bar(stacked=True, ax=ax, width=0.9, color=sns.color_palette('Set3'))
ax.set_ylim(0, 100)
# ax.set_title('Dominant Diabetic Drugs', weight='bold')
ax.set_xticklabels(['No Antibiotic', 'Q1', 'Q2', 'Q3', 'Q4'], rotation=0)
