==> 09_building_person_table.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path
import os
from statsmodels.stats.proportion import proportion_confint
from itertools import product

## Importing Data

person_df = pd.read_pickle('../../edited_fund_person_data.pkl')
person_df

province_id = pd.read_excel('../../diabeteschronicabuserprojectihio/Test_sample/Test/province_id.xlsx').set_index('province_id_ihio')
province_id.head()

## Building Empty Table

person_df['age_cat'].cat.categories

province_id.loc[[23, 27, 9, 10]]

idx_list = []
idx_list.append(pd.MultiIndex.from_product([['age_cat'], person_df['age_cat'].cat.categories]))
idx_list.append(pd.MultiIndex.from_product([['gender'], person_df['gender'].unique()]))
idx_list.append(pd.MultiIndex.from_product([['fund'], person_df['fund'].value_counts(ascending=False).index]))
idx_list.append(pd.MultiIndex.from_product([['province'], person_df['province'].value_counts(ascending=False).index]))
idx_list.append(pd.MultiIndex.from_tuples([('total', 'number')]))
idx_list

m_idx = []
for temp in idx_list:
    m_idx += temp.to_list()

cols = ['all'] + sorted(list(person_df['cat'].unique()))

table = pd.DataFrame(index=pd.MultiIndex.from_tuples(m_idx), columns=cols)
table

## Filling in the Values

def stringer(val, lower, upper, decimals=2):
    val = (val * 100).round(decimals)
    lower = (lower * 100).round(decimals)
    upper = (upper * 100).round(decimals)
    
    s = val.astype(str) + ' (' + lower.astype(str) + '-' + upper.astype(str) + ')'

    return s

st.t.interval(confidence=0.95,
             df=len(person_df)-1,
             loc=person_df['age'].mean(),
             scale=st.sem(person_df['age']))

metrics = ['age_cat', 'gender', 'fund', 'province', 'total']
cols

normalized_counts_index = pd.MultiIndex.from_tuples([(col, m, idx)for col, (m, idx) in product(cols, m_idx)])
normalized_counts = pd.DataFrame(
    index=normalized_counts_index, 
)
normalized_counts



for col in cols:
    for m in metrics:
    
        # print(col, m)

        if col == 'all':
            df = person_df
        else:
            df = person_df.loc[person_df['cat'] == col]
        
        if m == 'total':
            continue

        nobs = len(df)
        
        temp = df[m].value_counts(normalize=True)
        temp_count = df[m].value_counts()

        indexes = product([col], [m], temp.index)

        normalized_counts.loc[indexes, 'val'] = temp.values

        for idx in table.loc[m].index:

            lower, upper = proportion_confint(
                count=temp_count[idx],
                nobs=nobs,
                alpha=0.05,
                method='normal'
            )

            normalized_counts.loc[(col, m, idx), ['lower', 'upper']] = (lower, upper)
        


normalized_counts = normalized_counts.dropna()
s = stringer(normalized_counts['val'], normalized_counts['lower'], normalized_counts['upper'])
table = pd.DataFrame(s).unstack(0)[0]
table

temp = person_df['cat'].value_counts()

for col in cols:

    if col == 'all':
        table.loc[('total', 'number'), col] = len(person_df)
        continue
    
    table.loc[('total', 'number'), col] = temp[col]

temp

table

normalized_counts

## Saving person_table

# table.to_csv(r'E:\after_session_03_22\results\person_table.csv')
# table.to_clipboard()
