==> 09_trend_plot.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_csv('../../1403-4-16 export/trend_service_numbers.csv')
df

drug_class = ['J01A', 'J01B', 'J01C', 'J01D', 'J01E', 'J01F', 'J01G', 'J01M', 'J01X']

season_string = dict()
seasons = ['Spring', 'Summer', 'Fall', 'Winter']
for s in range(1, 13):
    year = 2014 + s // 4
    season = seasons[(s-1) % 4]
    season_string[s] = f'{year}-{season}'

season_string

color_list = []
for color in sns.color_palette('bright'):
    color_list.append(color)

def plot_ax_line(ax, data, drug):
    
    # Plotting the line
    sns.lineplot(data, ax=ax, x='season', y=drug, label=drug, lw=0.7, marker='.',
                 markersize=8, color=color_list[i])
    
    # Tweaking labels and titles
    ax.set_ylabel('')
    ax.set_title(f'Q{cat}')
    ax.set_xlabel('')
    ax.set_xticks(
        ticks=data['season'],
        labels=season_string.values(),
        rotation=45,
        ha='right'
    )
    ax.get_legend().set_visible(False)
    for axis in ['top','bottom','left','right']:
        ax.spines[axis].set_linewidth(0.1)

fig, axes = plt.subplots(2, 2, figsize=(10, 8), dpi=300, layout='constrained')
fig.set_facecolor('white')
# fig.suptitle('Number of Antibiotic Prescriptions per Drug Class', weight='bold')

for cat, ax in zip(sorted(df['cat'].unique()), axes.flatten()):
    for i, d in enumerate(drug_class):
        plot_ax_line(ax, df.loc[df['cat'] == cat], d)

leg_handels, leg_labels = axes[0, 0].get_legend_handles_labels()
fig.legend(
    leg_handels,
    leg_labels,
    loc='center',
    ncols=10,
    bbox_to_anchor=(0.5, -0.03)
)

## Data for values

df.replace({'season': season_string}, inplace=True)

guide_df = df.set_index(['cat', 'season']).unstack(0).swaplevel(0, 1, axis=1).sort_index(axis=1, level=0)

guide_df

# guide_df.to_excel('../trend_data_values.xlsx')
