==> 10_Building_prescription_table.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path
import os
from statsmodels.stats.proportion import proportion_confint
from itertools import product

## Importing Data and Building Data

parts_path = Path(r'E:\after_session_03_22\data\part_data')
os.listdir(parts_path)

## Building Empty Table

cols = ['all', 0, 1, 2, 3, 4]
table = pd.DataFrame(
    index=['ab', 'dm', 'total'],
    columns=cols,
)
table.fillna(0, inplace=True)
table

table.loc['ab', 2]

## Putting in the Values

for part_loc in os.listdir(parts_path):
    # Logging
    print('Processing:  ', part_loc)

    # Importing part
    part_loc = parts_path / part_loc
    part = pd.read_pickle(part_loc)

    # Filiing values
    categories = part['cat'].unique()
    for cat in categories:
        for drug in ['ab', 'dm', 'total']:
            cat_filt = part['cat'] == cat
            
            if drug == 'total':
                table.loc[drug, cat] = (table.loc[drug, cat]
                                        + part.loc[cat_filt].shape[0]
                                    )
                continue

            table.loc[drug, cat] = (table.loc[drug, cat]
                                    + part.loc[cat_filt].groupby('prescription')[drug].any().sum()
                                )

table

# Computing all column.
table['all'] = table.sum(axis=1)

table

## Saving

# Saving prescription_table_numbers
# It contains the number of prescriptions that have at least on ab or one dm drug in them by ab_groups.
# table.to_csv(r'E:\after_session_03_22\results\prescription_table_numbers.csv')

## Sandbox

part

part.groupby('prescription')['dm'].any()

part.groupby('prescription')['ab'].any()
