==> 11_editing_fund.ipynb <==
import pandas as pd
import numpy as np

prof['fund'] = prof['fund'].replace({4:0, 1:1, 2:2, 9:2, 11:3, 12:4, 13:4, 3:5, 15:5})
#0 = rural
#1 = civil servants
#2 = Iranian (+ khisfarma)
#3 = Universal
#4 = forigen
#5 = others

fund_dict = {4:0, 1:1, 2:2, 9:2, 11:3, 12:4, 13:4, 3:5, 15:5}
fund_name_dict = {
    0: 'Rural',
    1: 'Civil servants',
    2: 'Iranian + Self-employed',
    3: 'Universal',
    4: 'Foreign',
    5: 'Others'
}

# Person Table

df = pd.read_pickle('../../1403-4-16 export/yearly_costs_encr.pkl')['data']
df.fillna({'fund':4}, inplace=True)
df['fund'] = df['fund'].astype(int)
df

df['fund'].value_counts(normalize=True)

edited_fund_person_df = df.replace({'fund': fund_dict}).replace({'fund': fund_name_dict})

edited_fund_person_df

# edited_fund_person_df.to_pickle('edited_fund_person_data.pkl')
