==> 11_modifications_J_.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
from pathlib import Path

## Deleting J02, J04, J05

# Importing
one_df_days = pd.read_pickle(r'C:\Users\a.bagherian\Desktop\working\one_df_days\one_df_days_18_95.pickle')

one_df_days

# Importing service
service = pd.read_pickle(r'C:\Users\a.bagherian\Desktop\IHIO_share_A\service.pkl')
service.set_index('service', inplace=True)

atc_to_delete = ['J02', 'J04', 'J05']
service_to_delete = service.loc[service['atc_2'].isin(atc_to_delete)].index

one_df_days['service'].isin(service_to_delete).sum() / one_df_days.shape[0]

one_df_days_J01 = one_df_days.loc[~ one_df_days['service'].isin(service_to_delete)].copy()

1 - (one_df_days_J01['days'] / pd.Timedelta(days=1)).sum() / (
    one_df_days['days'] / pd.Timedelta(days=1)).sum()

# Saving one_df_days_J01
# one_df_days_J01.to_pickle(r'E:\after_session_03_22\data\one_df\one_df_days_18_95_J01.pkl')

## Building member_days_18_95_J01

one_df_days_J01

### Removing outliers

# Removing outliers
limit = pd.Timedelta(30, 'day')
print(limit)
outliers = one_df_days_J01.loc[one_df_days_J01['days'] > limit]
outliers

one_df_days_J01_outliers_removed = one_df_days_J01.loc[one_df_days_J01['days'] <= limit].copy()

one_df_days_J01_outliers_removed.describe()

member_days_J01 = one_df_days_J01_outliers_removed.groupby('member')['days'].sum()

member_days_J01.describe() / 3

# Saving member_days_18_95_outliers_removed_J01
# member_days_J01.to_pickle(r'E:\after_session_03_22\data\one_df\member_days_18_95_outliers_removed_J01.pkl')

### Building member_cat_J01

days_final = pd.read_pickle(r'E:\after_session_03_22\data\one_df\member_days_18_95_outliers_removed_J01.pkl')

days_final = days_final.to_frame()
days_final

days_final.loc[days_final['days'] > pd.Timedelta(0, 'day')].describe() / 3

num_days = (days_final['days'] / pd.Timedelta(1, 'day')) / 3
num_days 

days_final.loc[num_days == 0, 'cat'] = 0
days_final.loc[(0 < num_days) & (num_days <= 5), 'cat'] = 1
days_final.loc[(5 < num_days) & (num_days <= 11), 'cat'] = 2
days_final.loc[(11 < num_days) & (num_days <= 21), 'cat'] = 3
days_final.loc[(21 < num_days), 'cat'] = 4

days_final

days_final['cat'].value_counts(normalize=True, dropna=False)

# saving member_cat_J01
# days_final.to_pickle(r'E:\after_session_03_22\data\one_df\member_cat_J01.pkl')

## Building person_df_J01

person_df = pd.read_pickle(r'E:\after_session_03_22\data\person_df_3agegroups.pkl')
person_df

person_df.drop('cat', axis=1, inplace=True)

# Importing member_cat_J01
member_cat_J01 = pd.read_pickle(r'E:\after_session_03_22\data\one_df\member_cat_J01.pkl')

member_cat_J01

person_df_J01 = pd.merge(
    person_df,
    member_cat_J01[['cat']],
    how='left',
    left_index=True,
    right_index=True
)

person_df_J01['cat'].value_counts(normalize=True, dropna=False)

person_df_J01['cat'].fillna(0, inplace=True)

person_df_J01['cat'] = person_df_J01['cat'].astype(int)

person_df_J01

# Saving person_df_J01
# person_df_J01.to_pickle(r'E:\after_session_03_22\data\one_df\person_df_J01.pkl')

## Building parts with new ab column just based on J01

ab_service = service.loc[service['atc_2'] == 'J01'].index

ab_service

prev_parts_path = Path(r'C:\Users\a.bagherian\Desktop\working\Diabetics Combined 18_95')
os.listdir(prev_parts_path)

# Applying new ab columns and saving

# for part_num, prev_loc in enumerate(os.listdir(prev_parts_path)):
#     prev_loc = prev_parts_path / prev_loc
#     part_num += 1

#     # Importing previous part
#     part = pd.read_pickle(prev_loc)

#     # Modifying ab column to only consider J01
#     part['ab'] = part['service'].isin(ab_service)

#     # Modifying days column to only consider J01
#     part.loc[~ part['ab'], 'days'] = pd.Timedelta(0, 'day')

#     # Merging data from person_df_J01
#     part = pd.merge(
#         part,
#         person_df_J01.loc[
#             person_df_J01['part'] == part_num, ['fund', 'age', 'age_cat', 'cat']],
#         how='left',
#         left_on='member',
#         right_index=True
#     )
    
#     # Saving new parts
#     new_part_loc = (
#         Path(r'E:\after_session_03_22\data\part_data') / (prev_loc.stem + '_J01' + prev_loc.suffix)
#     )
#     part.to_pickle(new_part_loc)

part

## Replacing mod of province with province of parts

# Importing person_df_J01
person_df_J01 = pd.read_pickle(r'E:\after_session_03_22\data\one_df\person_df_J01.pkl')
person_df_J01

parts_path = Path(r'E:\after_session_03_22\data\part_data')
os.listdir(parts_path)

# Adding prov_mod column to parts

person_df_J01.rename(columns={'province': 'prov_mod'}, inplace=True)

for part_num, part_loc in enumerate(os.listdir(parts_path)):
    part_loc = parts_path / part_loc
    part_num += 1

    # Importing previous part
    part = pd.read_pickle(part_loc)

    # Adding prov_mod
    part = pd.merge(
        part,
        person_df_J01.loc[
            person_df_J01['part'] == part_num, ['prov_mod']
        ],
        how='left',
        left_on='member',
        right_index=True
    )

    # Saving new parts
    new_part_loc = (
        Path(r'E:\after_session_03_22\data\part_data\prov_mod') / part_loc.name
    )
    part.to_pickle(new_part_loc)
