==> 12_dm_drugclass.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path
import os
from statsmodels.stats.proportion import proportion_confint
from itertools import product

## Importing

parts_path = Path(r'E:\after_session_03_22\data\part_data')
os.listdir(parts_path)

## Calculating

# to_groupby = ['gender', 'age_cat', 'cat', 'member']
to_groupby = 'member'

first_iter = True
for part_loc in os.listdir(parts_path):
    # Logging
    print('Processing:  ', part_loc)

    # Importing part
    part_loc = parts_path / part_loc
    part = pd.read_pickle(part_loc)

    # Filtering for dm drugs
    part = part.loc[part['dm']]

    # Building dm_class column
    # Insulin services filter
    is_insulin = part['atc'].str[:4] == 'A10A'
    #
    part.loc[is_insulin, 'dm_class'] = 'A10A'
    part.loc[~ is_insulin, 'dm_class'] = part.loc[~ is_insulin, 'atc'].str[:5]
    # groupby
    if first_iter:
        df = part.groupby(to_groupby)['dm_class'].value_counts().to_frame().unstack(-1)
    first_iter = False
    df = pd.concat([
        df, part.groupby(to_groupby)['dm_class'].value_counts().to_frame().unstack(-1),
    ], axis=0)
    

# There seems to be some duplicates and are the same.
df = df.groupby(df.index).last()

df

part = pd.read_pickle(parts_path / 'part1_dm_combined_18_95_J01.pkl')

part['member'].nunique()

part.groupby('member')['dm_class'].value_counts().to_frame().unstack(-1)

## Adding Data From person_df_J01

person_df_J01 = pd.read_pickle(r'E:\after_session_03_22\data\one_df\person_df_J01.pkl')
person_df_J01

df.loc[:, 'dm_class']

# Merging
cols_to_merge = ['gender', 'fund', 'age_cat', 'cat', 'province']

final = pd.merge(
    df.loc[:, 'dm_class'],
    person_df_J01[cols_to_merge],
    how='left',
    on='member'
)

final

## Saving

# Saving dm_class_member data
# This file contains members and their gender and cat and age_cat and the number of their dm_drug_class services.
# For some reason the number of rows is more than
# final.to_pickle(r'E:\after_session_03_22\data\drug_class\dm_drug_class_numbers.pkl')
