==> 16_ab_service_trend.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path
import os
from statsmodels.stats.proportion import proportion_confint
from itertools import product

# Importing

parts_path = Path(r'E:\after_session_03_22\data\part_data')
os.listdir(parts_path)

## Defining

start_date = {
    1: '2014-03-20',
    2: '2014-06-21',
    3: '2014-09-23',
    4: '2014-12-22',
    5: '2015-03-21',
    6: '2015-06-21',
    7: '2015-09-23',
    8: '2015-12-22',
    9: '2016-03-20',
    10: '2016-06-21',
    11: '2016-09-22',
    12: '2016-12-21',
}
end_date = {
    1: '2014-06-20',
    2: '2014-09-22',
    3: '2014-12-21',
    4: '2015-03-20',
    5: '2015-06-20',
    6: '2015-09-22',
    7: '2015-12-21',
    8: '2016-03-19',
    9: '2016-06-20',
    10: '2016-09-21',
    11: '2016-12-20',
    12: '2017-03-20',
}

# Calculating

to_groupby = ['member', 'season']

first_iter = True
for part_loc in os.listdir(parts_path):
    # Logging
    print('Processing:  ', part_loc)

    # Importing part
    part_loc = parts_path / part_loc
    part = pd.read_pickle(part_loc)

    # Filtering for ab drugs
    part = part.loc[part['ab']]

    # Building ab_class column
    part['ab_class'] = part['atc'].str[:4]

    # Adding season
    for t in range(1, 12 + 1):
        part.loc[
            (part['date'] >= start_date[t]) & (part['date'] <= end_date[t]),
            'season'
        ] = t
    part['season'] = part['season'].astype(int)

    # groupby
    if first_iter:
        df = part.groupby(to_groupby)['ab_class'].value_counts().to_frame().unstack(-1)
    first_iter = False
    df = pd.concat([
        df, part.groupby(to_groupby)['ab_class'].value_counts().to_frame().unstack(-1)
    ])


df

## Adding cat from person_df_J01

person_df_J01 = pd.read_pickle(r'E:\after_session_03_22\data\one_df\person_df_J01.pkl')
person_df_J01

# Merging
cols_to_merge = ['cat']

ab_class_trend = pd.merge(
    df['ab_class'].reset_index('season'),
    person_df_J01[cols_to_merge],
    how='left',
    on='member'
)

ab_class_trend

# Dropping cat=0 because they where outliers.
outliers = (ab_class_trend.loc[ab_class_trend['cat'] == 0]).index
ab_class_trend.drop(outliers, axis=0, inplace=True)

## Saving

# Saving ab_class_trend
# ab_class_trend.to_pickle(r'E:\after_session_03_22\data\drug_class\ab_drug_class_numbers_seasonal.pkl')

## Building table for plotting

ab_class_trend = pd.read_pickle(r'E:\after_session_03_22\data\drug_class\ab_drug_class_numbers_seasonal.pkl')

ab_class_trend

ab_classes = ['J01A', 'J01B', 'J01C', 'J01D', 'J01E', 'J01F', 'J01G', 'J01M', 'J01X']

def table_builder(df):
    ab_table = pd.DataFrame(columns=)

trend_service_numbers = ab_class_trend.groupby(['cat', 'season'])[ab_classes].sum()

trend_service_numbers

## Saving

# Saving trend_service_numbers
# trend_service_numbers.to_csv(r'E:\after_session_03_22\results\trend_service_numbers.csv')
